fix: correct isLeftHandSide for non-update PrefixUnary and merge extras files
- isLeftHandSide previously returned true for every PrefixUnaryExpression
  parent. Upstream's helper only returns true for UpdateExpression (++ / --).
  Differentiate by operator so !x / ~x / +x / -x return false, matching
  upstream's ASI hazard semantics. Latent bug — no current test fires the
  needsLeadingSemicolon path with a `!`-parent + paren-leading replacement
  text, but the divergence would surface if future suggestion variants
  produce paren-leading code in `!` operand position.
- Consolidate the 5 _extras_*_test.go files (branches / fixer / options /
  realuser + general) into a single _extras_test.go with section banners.
  Same 252 cases, single Test function, simpler grep target.
- Add intentional-divergence comment to isStrongInnerNode explaining why
  NonNullExpression / ThisKeyword / SuperKeyword are added on top of
  utils.IsStrongPrecedenceNode.
